Polishing
Many things made better, like america

+++ Assets/Scripts/Effects/DelayedEffect.cs
@@ -7,7 +7,7 @@ public class DelayedEffect : MonoBehaviour
     private DamageCollider damageCollider;
 
     [SerializeField]private ParticleSystem particleEffect;
-    [SerializeField] private int lifetime=3;
+    [SerializeField] private float lifetime=3;
     private float startTime;
     private bool damageColliderEnabled = false;
